fixing delete workload for unsupported data services

+++ tests/pds/dataservice_test.go
@@ -905,12 +905,14 @@ var _ = Describe("{RestartPXPods}", func() {
 
 				defer func() {
 					Step("Delete the workload generating deployments", func() {
-						if ds.Name == "Cassandra" || ds.Name == "PostgreSQL" {
-							err = pdslib.DeleteK8sDeployments(dep.Name, namespace)
-						} else {
-							err = pdslib.DeleteK8sPods(pod.Name, namespace)
+						if !(ds.Name == mysql || ds.Name == kafka || ds.Name == zookeeper) {
+							if ds.Name == "Cassandra" || ds.Name == "PostgreSQL" {
+								err = pdslib.DeleteK8sDeployments(dep.Name, namespace)
+							} else {
+								err = pdslib.DeleteK8sPods(pod.Name, namespace)
+							}
+							Expect(err).NotTo(HaveOccurred())
 						}
-						Expect(err).NotTo(HaveOccurred())
 					})
 				}()
